+++ Service/Email/PurchaseCreditsEmail.php
@@ -15,7 +15,6 @@ use Symfony\Component\HttpFoundation\RequestStack;
 use Twig_Environment;
 use c975L\EmailBundle\Service\EmailServiceInterface;
 use c975L\PaymentBundle\Entity\Payment;
-use c975L\PaymentBundle\Entity\Payment;
 use c975L\ServicesBundle\Service\ServicePdfInterface;
 use c975L\PurchaseCreditsBundle\Service\Email\PurchaseCreditsEmailInterface;